requirements.txt
alembic==0.9.8
aniso8601==3.0.0
bcrypt==3.2.2
cffi==1.15.0
click==6.7
Flask==0.12.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.1.1
flask-restplus==0.10.1
flask-restx==0.5.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.7.1
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
Mako==1.3.5
MarkupSafe==2.0.1
pycparser==2.18
PyJWT==1.6.0
python-dateutil==2.7.0
python-editor==1.0.3
pytz==2018.3
six==1.11.0
SQLAlchemy==1.4.39
Werkzeug==0.14.1
psycopg2-binary==2.9.3
numpy
boto3==1.24.25
botocore~=1.27.96
Flask-JWT-Extended
scikit-learn==1.3.2